+++ src/Acr.UserDialogs.Android/Builders/TimePromptBuilder.cs
@@ -12,8 +12,7 @@ namespace Acr.UserDialogs.Builders
         public static Dialog Build(Activity activity, TimePromptConfig config)
         {
             var picker = new TimePicker(activity);
-            var builder = new AlertDialog
-                .Builder(activity)
+            var builder = new AlertDialog.Builder(activity) // TODO: style ID
                 .SetCancelable(false)
                 .SetTitle(config.Title)
                 .SetView(picker);