+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## 6.3.4
 * [fix][android] make sure to use appcompat activity dialog rendering in all cases
+* [fix][android] alert dialog was not using android styleID
 
 ## 6.3.3
 * [fix][droid] Fix NRE on droid 4.2.2 with soft keyboard dismiss on prompts